feat: re org layout so that all pages get navbar
