+++ apps/trading/client-pages/portfolio/portfolio.tsx
@@ -53,7 +53,7 @@ export const Portfolio = () => {
 
   const onMarketClick = useMarketClickHandler(true);
   const [sizes, handleOnLayoutChange] = usePaneLayout({ id: 'portfolio' });
-  const wrapperClasses = 'pt-0.5 h-full max-h-full flex flex-col';
+  const wrapperClasses = 'py-0.5 h-full max-h-full flex flex-col';
   return (
     <div className={wrapperClasses}>
       <ResizableGrid vertical onChange={handleOnLayoutChange}>
