--- apps/trading/components/layout/index.ts
@@ -1,1 +1,0 @@
-export * from './layout';
